+++ src/main/php/broker/param/SecretParamBroker.php
@@ -12,22 +12,22 @@ use stubbles\input\filter\range\StringLength;
 use stubbles\input\valuereader\CommonValueReader;
 use stubbles\lang\reflect\annotation\Annotation;
 /**
- * Filter boolean values based on a @Request[SecureString] annotation.
+ * Filter boolean values based on a @Request[Secret] annotation.
  *
  * @since  3.0.0
  */
-class SecureStringParamBroker extends MultipleSourceParamBroker
+class SecretParamBroker extends MultipleSourceParamBroker
 {
     /**
      * handles single param
      *
      * @param   \stubbles\input\valuereader\CommonValueReader  $valueReader  instance to filter value with
      * @param   \stubbles\lang\reflect\annotation\Annotation   $annotation   annotation which contains filter metadata
-     * @return  \stubbles\lang\SecureString
+     * @return  \stubbles\lang\Secret
      */
     protected function filter(CommonValueReader $valueReader, Annotation $annotation)
     {
-        return $valueReader->asSecureString(
+        return $valueReader->asSecret(
                 new StringLength($annotation->getMinLength(),
                                  $annotation->getMaxLength()
                 )