+++ src/main/php/valuereader/CommonValueReader.php
@@ -66,9 +66,18 @@ interface CommonValueReader
      * read as string value
      *
      * @param   \stubbles\input\filter\range\StringLength  $length
-     * @return  \stubbles\lang\SecureString
+     * @return  \stubbles\lang\Secret
+     * @deprecated  since 6.0.0, use asSecret() instead, will be removed with 7.0.0
      */
     public function asSecureString(StringLength $length = null);
+
+    /**
+     * read as secret
+     *
+     * @param   \stubbles\input\filter\range\StringLength  $length
+     * @return  \stubbles\lang\Secret
+     */
+    public function asSecret(StringLength $length = null);
 
     /**
      * read as text value
@@ -90,7 +99,7 @@ interface CommonValueReader
      * read as password value
      *
      * @param   \stubbles\input\filter\PasswordChecker  $checker  checker to be used to ensure a good password
-     * @return  \stubbles\lang\SecureString
+     * @return  \stubbles\lang\Secret
      */
     public function asPassword(PasswordChecker $checker);