+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### BC breaks
 
   * Raised minimum required PHP version to 5.5
+  * deprecated `stubbles\input\ValueReader::asSecureString()`, use `stubbles\input\ValueReader::asSecret()` instead, will be removed with 7.0.0
+  * deprecated `@Request[SecureString]`, use `@Request[Secret]` instead, will be removed with 7.0.0
 
 
 ### Other changes